error/log: introduce error wrapping
Introduce error wrapping to carry trace information across the multiple
layers. Instead of encoding the trace in the error message we set it
into a separate field, this is used for the logging system when dealing
with error specifically, this wrapping will be useful for UI interactions
to split between "debugging" information and actual user warnings and
messages.

The log subsystem also brings a log level validation as well as a new
Out() facility to differentiate commands output.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>