network: test use curl instead of swupd
Instead of using swupd to test the network use a curl call. With that
we don't hang until the swupd cache download (letting the user wait
for a unnecessarily long operation).

We use a curl call (instead of a native call) in order to validate the
proxy configuration and to abstract any possible protocol behind the
version url.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>