storage: implement IsAvailable()/LoadAvailableBlockDevices() methods
IsAvailable() method will return true if the storage device was in use
when we first loaded its description. Relying only on state flag is not
fairly accurate depending on the hardware configuration - i.e if the
storage is not external and has an OS loaded from.

LoadAvailableBlockDevice() uses the available flag and return only
block devices marked as available.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>