storage: use same objects across manual and guided
This patch changes the storage subsystem and TUI to operate on the
same objects while doing manual or guided. The TUI code keeps a copy
of the BlockDevice and operate on it while in the respective screens
and then commits to the BlockDevice subsystem and keep the pointer
to the selected block device into the data model.

This patch also fixes an issue found by Caio that when you do a guided
partitioning and then tries to do a manual partitioning the resulting
selected block device points to the one selected with the guided and
the changes done with the manual partitioning will never be committed
to the data model.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>